djangobackend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.2
defusedxml==0.7.1
distro==1.9.0
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-ses==4.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
Markdown==3.8
MarkupSafe==3.0.2
numpy==2.2.5
oauthlib==3.2.2
openai==1.76.0
packaging==25.0
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.13.0
sendgrid==6.12.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==2.3.8
